+++ src/gql/__tests__/assembleGqlContent/multiple_services/output.js
@@ -1,7 +1,7 @@
 import gql from 'graphql-tag'
 
 export const buildRequest_getConfig = gql`
-  query buildRequest_getConfig($req: GetConfigRequestInput) {
+  query buildRequest_getConfig($req: GetConfigRequest_Input) {
     buildRequest_getConfig(req: $req) {
       name
       path
@@ -10,7 +10,7 @@ export const buildRequest_getConfig = gql`
 `
 
 export const anotherRequest_getConfig = gql`
-  query anotherRequest_getConfig($req: GetConfigRequestInput) {
+  query anotherRequest_getConfig($req: GetConfigRequest_Input) {
     anotherRequest_getConfig(req: $req) {
       name
       path